+++ openmls/src/treesync/node/leaf_node.rs
@@ -164,7 +164,6 @@ impl LeafNodeParametersBuilder {
 ///     opaque signature<V>;
 /// } LeafNode;
 /// ```
-// TODO(#1242): Do not derive `TlsDeserialize`.
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, TlsSerialize, TlsSize)]
 pub struct LeafNode {
     payload: LeafNodePayload,
@@ -471,7 +470,7 @@ impl LeafNode {
     /// Perform all checks that can be done without further context:
     /// - the used extensions are not known to be invalid in leaf nodes
     /// - the types of the used extensions are covered by the capabilities
-    /// - the type of the credential is coveered by the capabilities
+    /// - the type of the credential is covered by the capabilities
     pub(crate) fn validate_locally(&self) -> Result<(), LeafNodeValidationError> {
         // Check that no extension is invalid when used in leaf nodes.
         let invalid_extension_types = self
@@ -500,6 +499,7 @@ impl LeafNode {
         }
 
         // Check that the capabilities contain the leaf node's credential type.
+        // (https://validation.openmls.tech/#valn0113)
         if !self
             .capabilities()
             .contains_credential(self.credential().credential_type())